Refactor OSS `DSv2Incompatible` and `InMemorySparkTable` trait structure
