+++ spark/src/test/scala/org/apache/spark/sql/delta/DeltaTestUtils.scala
@@ -608,6 +608,75 @@ trait DeltaTestUtilsForTempViews
   }
 }
 
+trait DeltaSQLInMemoryTestUtils
+    extends DeltaSQLTestUtils
+    with InMemoryTestTableMixin {
+
+  import org.apache.spark.sql.delta.catalog.InMemoryDeltaCatalog
+
+  /**
+   * Override for [[withTable]] that asserts no leftover physical parquet as a sanity-check
+   * for V2 paths.
+   */
+  override protected def withTable(tableNames: String*)(f: => Unit): Unit = {
+    try {
+      super.withTable(tableNames: _*) {
+        f
+        tableNames.foreach(assertNoParquetFiles)
+      }
+    } finally {
+      for (tableName <- tableNames) {
+        assert(!InMemoryDeltaCatalog.contains(tableName))
+      }
+    }
+  }
+
+  /**
+   * Overrides for [[afterEach]], cleans the [[InMemoryDeltaCatalog]] after each test.
+   */
+  override protected def afterEach(): Unit = {
+    try {
+      InMemoryDeltaCatalog.reset()
+    } finally {
+      super.afterEach()
+    }
+  }
+
+  /**
+   * Assert no parquet files are contain within the data directory for [[tableName]].
+   * Used to sanity-check our V2-only write paths.
+   */
+  protected def assertNoParquetFiles(tableName: String): Unit = {
+    import java.nio.file.{Files, Path}
+
+    val ident = TableIdentifier(tableName)
+
+    // Temp views don't create anything on the disk, but still use `withTable`.
+    // Either way, something that doesn't have a catalog entry, but is used in `withTable` should
+    // not be checked on disk.
+    if (!spark.sessionState.catalog.tableExists(ident)) {
+      return
+    }
+
+    val catalogTable = spark.sessionState.catalog.getTableMetadata(ident)
+    val dataPath = new File(new java.net.URI(catalogTable.location.toString))
+    if (dataPath.exists()) {
+      val stream = Files.walk(dataPath.toPath)
+      try {
+        val parquetFiles = stream
+          .filter(Files.isRegularFile(_))
+          .filter(_.toString.endsWith(".parquet"))
+          .toArray.map(_.asInstanceOf[Path].toString).toSeq
+        assert(parquetFiles.isEmpty,
+          s"Physical parquet files found while V2 in-memory mode is enabled. " +
+          s"DML may have fallen back to V1. Files: $parquetFiles")
+      } finally {
+        stream.close()
+      }
+    }
+  }
+}
+
 /**
  * Trait collecting helper methods for DML tests e.p. creating a test table for each test and
  * cleaning it up after each test.
@@ -755,6 +824,31 @@ trait DeltaDMLTestUtils
     )
       .drop(CDCReader.CDC_COMMIT_TIMESTAMP)
       .drop(CDCReader.CDC_COMMIT_VERSION)
+  }
+}
+
+/**
+ * Overrides [[DeltaDMLTestUtils]] methods to route writes through the V2 in-memory path.
+ */
+trait DeltaDMLInMemoryTestUtils
+    extends DeltaDMLTestUtils
+    with DeltaSQLInMemoryTestUtils {
+
+  /**
+   * Appends [[df]] into the test table.
+   */
+  override protected def append(df: DataFrame, partitionBy: Seq[String] = Nil): Unit = {
+    if (!spark.sessionState.catalog.tableExists(tableIdentifier)) {
+      val partitioning = if (partitionBy.nonEmpty) {
+        s"PARTITIONED BY (${partitionBy.mkString(", ")})"
+      } else {
+        ""
+      }
+      spark.sql(
+        s"CREATE TABLE $tableSQLIdentifier (${df.schema.toDDL}) USING delta $partitioning")
+    }
+    df.writeTo(tableSQLIdentifier).append()
+    assertNoParquetFiles(tableSQLIdentifier)
   }
 }
